// src/generated/wotlk/apl.ts
export interface APLRotation {
    /**
     * @generated from protobuf field: proto.APLRotation.Type type = 3
     */
    type: APLRotation_Type;
    /**
     * @generated from protobuf field: proto.SimpleRotation simple = 4
     */
    simple?: SimpleRotation;
    /**
     * @generated from protobuf field: repeated proto.APLPrepullAction prepull_actions = 1
     */
    prepullActions: APLPrepullAction[];
    /**
     * @generated from protobuf field: repeated proto.APLListItem priority_list = 2
     */
    priorityList: APLListItem[];
}
/**
 * @generated from protobuf enum proto.APLRotation.Type
 */
export enum APLRotation_Type {
    /**
     * @generated from protobuf enum value: TypeUnknown = 0;
     */
    TypeUnknown = 0,
    /**
     * @generated from protobuf enum value: TypeAuto = 1;
     */
    TypeAuto = 1,
    /**
     * @generated from protobuf enum value: TypeSimple = 2;
     */
    TypeSimple = 2,
    /**
     * @generated from protobuf enum value: TypeAPL = 3;
     */
    TypeAPL = 3,
    /**
     * @generated from protobuf enum value: TypeLegacy = 4;
     */
    TypeLegacy = 4
}
/**
 * @generated from protobuf message proto.SimpleRotation
 */
export interface SimpleRotation {
    /**
     * @generated from protobuf field: string spec_rotation_json = 1
     */
    specRotationJson: string;
    /**
     * @generated from protobuf field: proto.Cooldowns cooldowns = 2
     */
    cooldowns?: Cooldowns;
}
/**
 * @generated from protobuf message proto.APLPrepullAction
 */
export interface APLPrepullAction {
    /**
     * @generated from protobuf field: proto.APLAction action = 1
     */
    action?: APLAction;
    /**
     * @generated from protobuf field: proto.APLValue do_at_value = 4
     */
    doAtValue?: APLValue; // When to perform this prepull action. Should be a negative value.
    /**
     * @generated from protobuf field: bool hide = 3
     */
    hide: boolean; // Causes this item to be ignored.
}
/**
 * @generated from protobuf message proto.APLListItem
 */
export interface APLListItem {
    /**
     * @generated from protobuf field: bool hide = 1
     */
    hide: boolean; // Causes this item to be ignored.
    /**
     * @generated from protobuf field: string notes = 2
     */
    notes: string; // Comments for the reader.
    /**
     * @generated from protobuf field: proto.APLAction action = 3
     */
    action?: APLAction; // The action to be performed.
}
// ...
class APLRotation$Type extends MessageType<APLRotation> {
    constructor() {
        super("proto.APLRotation", [
            { no: 3, name: "type", kind: "enum", T: () => ["proto.APLRotation.Type", APLRotation_Type] },
            { no: 4, name: "simple", kind: "message", T: () => SimpleRotation },
            { no: 1, name: "prepull_actions", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => APLPrepullAction },
            { no: 2, name: "priority_list", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => APLListItem }
        ]);
    }
    create(value?: PartialMessage<APLRotation>): APLRotation {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.type = 0;
        message.prepullActions = [];
        message.priorityList = [];
        if (value !== undefined)
            reflectionMergePartial<APLRotation>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: APLRotation): APLRotation {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* proto.APLRotation.Type type */ 3:
                    message.type = reader.int32();
                    break;
                case /* proto.SimpleRotation simple */ 4:
                    message.simple = SimpleRotation.internalBinaryRead(reader, reader.uint32(), options, message.simple);
                    break;
                case /* repeated proto.APLPrepullAction prepull_actions */ 1:
                    message.prepullActions.push(APLPrepullAction.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* repeated proto.APLListItem priority_list */ 2:
                    message.priorityList.push(APLListItem.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: APLRotation, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated proto.APLPrepullAction prepull_actions = 1; */
        for (let i = 0; i < message.prepullActions.length; i++)
            APLPrepullAction.internalBinaryWrite(message.prepullActions[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* repeated proto.APLListItem priority_list = 2; */
        for (let i = 0; i < message.priorityList.length; i++)
            APLListItem.internalBinaryWrite(message.priorityList[i], writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* proto.APLRotation.Type type = 3; */
        if (message.type !== 0)
            writer.tag(3, WireType.Varint).int32(message.type);
        /* proto.SimpleRotation simple = 4; */
        if (message.simple)
            SimpleRotation.internalBinaryWrite(message.simple, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message proto.APLRotation
 */
export const APLRotation = new APLRotation$Type();
// @generated message type with reflection information, may provide speed optimized methods
class SimpleRotation$Type extends MessageType<SimpleRotation> {
    constructor() {
        super("proto.SimpleRotation", [
            { no: 1, name: "spec_rotation_json", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "cooldowns", kind: "message", T: () => Cooldowns }
        ]);
    }
    create(value?: PartialMessage<SimpleRotation>): SimpleRotation {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.specRotationJson = "";
        if (value !== undefined)
            reflectionMergePartial<SimpleRotation>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SimpleRotation): SimpleRotation {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string spec_rotation_json */ 1:
                    message.specRotationJson = reader.string();
                    break;
                case /* proto.Cooldowns cooldowns */ 2:
                    message.cooldowns = Cooldowns.internalBinaryRead(reader, reader.uint32(), options, message.cooldowns);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SimpleRotation, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string spec_rotation_json = 1; */
        if (message.specRotationJson !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.specRotationJson);
        /* proto.Cooldowns cooldowns = 2; */
        if (message.cooldowns)
            Cooldowns.internalBinaryWrite(message.cooldowns, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message proto.SimpleRotation
 */
export const SimpleRotation = new SimpleRotation$Type();
// @generated message type with reflection information, may provide speed optimized methods
class APLPrepullAction$Type extends MessageType<APLPrepullAction> {
    constructor() {
        super("proto.APLPrepullAction", [
            { no: 1, name: "action", kind: "message", T: () => APLAction },
            { no: 4, name: "do_at_value", kind: "message", T: () => APLValue },
            { no: 3, name: "hide", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
    create(value?: PartialMessage<APLPrepullAction>): APLPrepullAction {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.hide = false;
        if (value !== undefined)
            reflectionMergePartial<APLPrepullAction>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: APLPrepullAction): APLPrepullAction {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* proto.APLAction action */ 1:
                    message.action = APLAction.internalBinaryRead(reader, reader.uint32(), options, message.action);
                    break;
                case /* proto.APLValue do_at_value */ 4:
                    message.doAtValue = APLValue.internalBinaryRead(reader, reader.uint32(), options, message.doAtValue);
                    break;
                case /* bool hide */ 3:
                    message.hide = reader.bool();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: APLPrepullAction, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* proto.APLAction action = 1; */
        if (message.action)
            APLAction.internalBinaryWrite(message.action, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* bool hide = 3; */
        if (message.hide !== false)
            writer.tag(3, WireType.Varint).bool(message.hide);
        /* proto.APLValue do_at_value = 4; */
        if (message.doAtValue)
            APLValue.internalBinaryWrite(message.doAtValue, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message proto.APLPrepullAction
 */
export const APLPrepullAction = new APLPrepullAction$Type();
// ...
export const APLListItem = new APLListItem$Type();
// ...
export const APLAction = new APLAction$Type();
// ...
export const APLValue = new APLValue$Type();